+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This changelog file was started on October 2019. Previous changes were more or l
 ### Added
 - Rtp bundle can be enabled per LinphoneAccount, superseeding the setting at LinphoneCore level.
 - New APIs on Friend object to be able to set more info such as a Picture, Organization, Native ID & Starred
+- QRCode image generation.
 
 ### Changed
 - Auto schedule of core.iterate() method now uses a higher delay for timer if app is in background